@@ -12,11 +12,11 @@
     "axios": "^0.24.0",
     "dotenv": "^10.0.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-bootstrap": "^2.0.3",
-    "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.0",
     "react-stripe-checkout": "^2.6.3",
     "redux": "^4.1.2",
