@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
+import { Routes, Route, Navigate } from "react-router-dom";
+import { useSelector } from "react-redux";
 import axios from "axios";
 
 import Home from "./pages/Home/Home";
@@ -10,11 +11,10 @@ import TopNav from "./components/TopNav/TopNav";
 import Footer from "./components/Footer/Footer";
 import Announcement from "./components/Announcement/Announcement";
 import ProductList from "./pages/ProductList/ProductList";
-import Product from "./components/Products/Product/Product";
+// import Product from "./components/Products/Product/Product";
 import SingleProduct from "./pages/SingleProduct/SingleProduct";
 import Pay from "./pages/Stripe/Pay";
 import Success from "./pages/Stripe/Success";
-import { useSelector } from "react-redux";
 import Account from "./pages/Account/Account";
 
 const App = () => {
@@ -26,7 +26,6 @@ const App = () => {
   useEffect(() => {
     const getSales = async () => {
       try {
-        // const res = await axios.get("/sales/");
         const res = await axios.get(
           process.env.REACT_APP_SERVER_URL + "/sales/"
         );
@@ -39,47 +38,49 @@ const App = () => {
   }, []);
 
   return (
-    <HashRouter>
+    <React.Fragment>
       {sales && <Announcement sales={sales} />}
       <TopNav currentUser={user} />
-      <Switch>
-        <Route exact path="/">
-          <Home sales={sales} />
-        </Route>
-        <Route path="/product-list/:category">
-          <ProductList sales={sales} />
-        </Route>
-        <Route path="/product:id">
-          <Product />
-        </Route>
-        <Route path="/product">
-          <SingleProduct sales={sales} />
-        </Route>
-        <Route path="/register">
-          {user ? <Redirect to="/" /> : <Register errorStatus={error} />}
-        </Route>
-        <Route path="/login">
-          {user ? <Redirect to="/" /> : <Login errorStatus={error} />}
-        </Route>
-        <Route path="/account/:id">
-          {!user ? (
-            <Redirect to="/login" />
-          ) : (
-            <Account currentUser={user} errorStatus={error} />
-          )}
-        </Route>
-        <Route path="/cart">
-          <Cart currentUserId={user?._id} sales={sales} />
-        </Route>
-        <Route path="/pay">
-          <Pay />
-        </Route>
-        <Route path="/success">
-          <Success />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route exact path="/" element={<Home sales={sales} />} />
+        <Route
+          path="/product-list/:category"
+          element={<ProductList sales={sales} />}
+        />
+        {/* <Route path="/product:productId" element={<Product />} /> */}
+        <Route
+          path="/product/:productId"
+          element={<SingleProduct sales={sales} />}
+        />
+        <Route
+          path="/register"
+          element={
+            user ? <Navigate to="/" /> : <Register errorStatus={error} />
+          }
+        />
+        <Route
+          path="/login"
+          element={user ? <Navigate to="/" /> : <Login errorStatus={error} />}
+        />
+        <Route
+          path="/account/:id"
+          element={
+            !user ? (
+              <Navigate to="/login" />
+            ) : (
+              <Account currentUser={user} errorStatus={error} />
+            )
+          }
+        />
+        <Route
+          path="/cart"
+          element={<Cart currentUserId={user?._id} sales={sales} />}
+        />
+        <Route path="/pay" element={<Pay />} />
+        <Route path="/success" element={<Success />} />
+      </Routes>
       <Footer currentUser={user} />
-    </HashRouter>
+    </React.Fragment>
   );
 };
 
